TPM is now part of FSW core
