+++ flight/apps/adcs/ad.py
@@ -16,8 +16,8 @@ from apps.adcs.math import R_to_quat, quat_to_R, quaternion_multiply, skew
 from apps.adcs.orbit_propagation import OrbitPropagator
 from apps.adcs.sun import SUN_VECTOR_STATUS, approx_sun_position_ECI, compute_body_sun_vector_from_lux, read_light_sensors
 from apps.telemetry.constants import GPS_IDX
-from apps.time_processor.time_processor import TimeProcessor
 from core import DataHandler as DH
+from core.time_processor import TimeProcessor
 from hal.configuration import SATELLITE
 from ulab import numpy as np
