+++ flight/core/data_handler.py
@@ -43,6 +43,7 @@ import struct
 import time
 
 from core.logging import logger
+from core.time_processor import TimeProcessor
 from micropython import const
 
 try:
@@ -309,7 +310,7 @@ class DataProcess:
             str: The new filename.
         """
         # Keeping the tag name in the filename for identification in debugging
-        return join_path(self.dir_path, self.tag_name) + "_" + str(int(time.time())) + ".bin"
+        return join_path(self.dir_path, self.tag_name) + "_" + str(int(TimeProcessor.time())) + ".bin"
 
     def try_to_reuse_latest_file(self) -> bool:
         """
@@ -609,7 +610,7 @@ class ImageProcess(DataProcess):
             str: The new filename.
         """
         # Keeping the tag name in the filename for identification
-        return join_path(self.dir_path, self.tag_name) + "_" + str(int(time.time())) + ".jpg"
+        return join_path(self.dir_path, self.tag_name) + "_" + str(int(TimeProcessor.time())) + ".jpg"
 
     def log(self, data: bytearray) -> None:
         """
