+++ flight/core/logging.py
@@ -52,6 +52,7 @@ import sys
 import time
 from collections import namedtuple
 
+from core.time_processor import TimeProcessor
 from micropython import const
 
 try:
@@ -174,7 +175,7 @@ class Formatter:
             "args": record.args,
         }
         if "{asctime}" in self.fmt or "%(asctime)s" in self.fmt:
-            now = time.localtime()
+            now = time.localtime(TimeProcessor.time())
             # pylint: disable=line-too-long
             vals[
                 "asctime"
