+++ flight/tasks/gps.py
@@ -1,7 +1,5 @@
 # GPS Task
 
-# import time
-
 from apps.telemetry.constants import GPS_IDX
 from core import DataHandler as DH
 from core import TemplateTask
